jarrut: modulaarisuutta ja lokaalia purkkahackia
viestit pc -> driver -> braker toimii, jarruservot pyörii

+++ carcomputer/comm.h
@@ -5,9 +5,9 @@
 // FIXME BUF_TXHOST into config etc
 #include "ringbuf.h"
 
-void dump_info(uint8_t stream, uint16_t type, uint16_t size, void *data);
-void recv_info(uint8_t stream, uint16_t *type, uint16_t *size, void *data);
-void recv_info_check(uint8_t stream, uint16_t type, uint16_t size, void *data);
+void dump_info(uint8_t stream, uint8_t type, uint8_t size, void *data);
+void recv_info(uint8_t stream, uint8_t *type, uint8_t *size, void *data);
+void recv_info_check(uint8_t stream, uint8_t type, uint8_t size, void *data);
 
 #define DUMP_INFO(type, data) dump_info(BUF_TXHOST, type, sizeof(data), &data)
